flask_board/templates/jinja2/celery/requirements.txt
flask
click
gunicorn
python-dotenv
celery
pytest
